Move rename ReservationAccessCode-component for future wrapping
